Add error types, stop returning strings
In Newsboat, we relay the error to the user, but other people might want
to do other things. This commit replaces `String`s with two new types,
one for `new()` and one for `matches()`. Even though they are identical,
I decided to have two of them because in the future they might become
slightly different. (I don't foresee that, as the underlying API has
been stable for decades, but if it ever happens I'd be glad I did it
this way.)

As a result, we can drop dependencies on `gettext-rs` and `strprintf`.